Update prediction
The Python script provided creates a new GitHub repository, adds a README.md file to it, and inserts a link to your <a href"https://ecogreenitrecycling.co.uk/computer-recycling/"> website </a>. Make sure to replace placeholders with your own GitHub access token, username, desired repository name, and website URL in the script. This allows you to easily create a GitHub repository with a backlink to your website using Python and the PyGitHub library.